Refactor propensity_score to outer utils
