+++ zepid/causal/ipw/IPCW.py
@@ -2,7 +2,7 @@ import warnings
 import numpy as np
 import pandas as pd
 
-from .utils import propensity_score
+from zepid.causal.utils import propensity_score
 
 
 class IPCW:
