+++ zepid/causal/ipw/__init__.py
@@ -1,4 +1,3 @@
-from .utils import propensity_score
 from .IPTW import IPTW
 from .IPMW import IPMW
 from .IPCW import IPCW
